requirements.txt
alabaster==0.7.10
anaconda-client==1.6.5
anaconda-navigator==1.6.9
anaconda-project==0.8.0
appdirs==1.4.3
asgiref==3.2.10
asn1crypto==0.22.0
astroid==1.5.3
astropy==2.0.2
Babel==2.5.0
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.0.0
bokeh==0.12.10
boto==2.48.0
Bottleneck==1.2.1
certifi==2017.7.27.1
cffi==1.10.0
chardet==3.0.4
click==6.7
cloudpickle==0.4.0
clyent==1.2.2
colorama==0.3.9
conda==4.5.12
conda-build==3.0.27
conda-verify==2.0.0
contextlib2==0.5.5
cryptography==2.0.3
cycler==0.10.0
Cython==0.26.1
cytoolz==0.8.2
dask==0.15.3
datashape==0.5.4
decorator==4.1.2
distlib==0.3.0
distributed==1.19.1
Django==3.1.2
docutils==0.14
entrypoints==0.2.3
et-xmlfile==1.0.1
fastcache==1.0.2
filelock==3.0.12
Flask==0.12.2
Flask-Cors==3.0.3
gcloud==0.17.0
gevent==1.2.2
glob2==0.5
gmpy2==2.0.8
googleapis-common-protos==1.52.0
greenlet==0.4.12
h5py==2.7.0
heapdict==1.0.0
html5lib==0.999999999
httplib2==0.18.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
importlib-metadata==1.6.0
importlib-resources==1.5.0
ipykernel==4.6.1
ipython==6.1.0
ipython-genutils==0.2.0
ipywidgets==7.0.0
isort==4.2.15
itsdangerous==0.24
jdcal==1.3
jedi==0.10.2
jeepney==0.4
Jinja2==2.9.6
jsonschema==2.6.0
jupyter-client==5.1.0
jupyter-console==5.2.0
jupyter-core==4.3.0
jupyterlab==0.27.0
jupyterlab-launcher==0.4.0
jws==0.1.3
keyring==17.0.0
lazy-object-proxy==1.3.1
llvmlite==0.20.0
locket==0.2.0
lxml==4.1.0
MarkupSafe==1.0
matplotlib==2.1.0
mccabe==0.6.1
mistune==0.7.4
mpmath==0.19
msgpack-python==0.4.8
multipledispatch==0.4.9
mysqlclient==2.0.1
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.0
nltk==3.2.4
nose==1.3.7
notebook==5.0.0
numba==0.35.0+10.g143f70e90
numexpr==2.6.2
numpy==1.13.3
numpydoc==0.7.0
oauth2client==3.0.0
odo==0.5.1
olefile==0.44
openpyxl==2.4.8
packaging==16.8
pandas==0.20.3
pandocfilters==1.4.2
partd==0.3.8
path.py==10.3.1
pathlib2==2.3.0
patsy==0.4.1
pep8==1.7.0
pexpect==4.2.1
pickleshare==0.7.4
Pillow==4.2.1
pkginfo==1.4.1
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.13.0
psutil==5.4.0
ptyprocess==0.5.2
py==1.4.34
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.4.3
pycurl==7.43.0
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.7.1
pylint==1.7.4
PyMySQL==0.10.1
pyodbc==4.0.17
pyOpenSSL==17.2.0
pyparsing==2.2.0
PyPDF2==1.26.0
Pyrebase==3.0.27
PySocks==1.6.7
pytest==3.2.1
python-dateutil==2.6.1
python-jwt==2.0.1
python-pptx==0.6.18
pytz==2017.2
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.5.2
requests==2.11.1
requests-toolbelt==0.7.0
rope==0.10.5
rsa==4.6
ruamel-yaml==0.11.14
scikit-image==0.13.0
scikit-learn==0.19.1
scipy==0.19.1
seaborn==0.8
SecretStorage==3.1.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.11.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.7
Sphinx==1.6.3
sphinxcontrib-websupport==1.0.1
spyder==3.3.2
spyder-kernels==0.2.4
SQLAlchemy==1.1.13
sqlparse==0.4.1
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tblib==1.3.2
terminado==0.6
testpath==0.3.1
toolz==0.8.2
tornado==4.5.2
traitlets==4.3.2
twilio==6.22.0
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
virtualenv==20.0.20
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.12.2
whitenoise==5.2.0
widgetsnbextension==3.0.2
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
zipp==3.1.0
